Wrap event properties for Element.
