+++ packages/shadydom/src/patches/Element.js
@@ -12,6 +12,7 @@ import * as utils from '../utils.js';
 import {scopeClassAttribute} from '../style-scoping.js';
 import {shadyDataForNode} from '../shady-data.js';
 import {attachShadow, ownerShadyRootForNode} from '../attach-shadow.js';
+import {eventPropertyNamesForElement, wrappedDescriptorForEventProperty} from '../patch-events.js';
 
 const doc = window.document;
 
@@ -121,6 +122,12 @@ export const ElementPatches = utils.getOwnPropertyDescriptors({
   }
 
 });
+
+if (!utils.settings.preferPerformance) {
+  eventPropertyNamesForElement.forEach(property => {
+    ElementPatches[property] = wrappedDescriptorForEventProperty(property);
+  });
+}
 
 export const ElementShadowPatches = utils.getOwnPropertyDescriptors({
   /**
