+++ packages/shadydom/src/wrapper.js
@@ -9,7 +9,7 @@ subject to an additional IP rights grant found at http://polymer.github.io/PATEN
 */
 
 import * as utils from './utils.js';
-import {eventPropertyNamesForHTMLElement} from './patch-events.js';
+import {eventPropertyNamesForElement, eventPropertyNamesForHTMLElement} from './patch-events.js';
 
 /** @implements {IWrapper} */
 class Wrapper {
@@ -247,7 +247,7 @@ class Wrapper {
 
 }
 
-eventPropertyNamesForHTMLElement.forEach(name => {
+const addEventPropertyWrapper = (name) => {
   Object.defineProperty(Wrapper.prototype, name, {
     /** @this {Wrapper} */
     get() {
@@ -259,8 +259,10 @@ eventPropertyNamesForHTMLElement.forEach(name => {
     },
     configurable: true
   });
+};
 
-});
+eventPropertyNamesForElement.forEach(name => addEventPropertyWrapper(name));
+eventPropertyNamesForHTMLElement.forEach(name => addEventPropertyWrapper(name));
 
 export {Wrapper};
